refactor(配置管理): 将配置文件初始化逻辑移至define模块
重构配置文件初始化逻辑，从run.py移动到define.py
移除未使用的pytest导入并清理测试文件空行
